Migrate bgfx examples
Moving all bgfx tests requiring assets and JOML to lwjgl-demos.
